+++ client/src/pages/_document.tsx
@@ -1,4 +1,6 @@
-import Document, { Html, Head, Main, NextScript } from 'next/document';
+import Document, {
+  Html, Head, Main, NextScript,
+} from 'next/document';
 
 export default class MyDocument extends Document {
   render() {
