+++ client/tsconfig.json
@@ -1,22 +1,10 @@
 {
   "compilerOptions": {
     "target": "es6",
-    "lib": [
-      "dom",
-      "dom.iterable",
-      "esnext"
-    ],
-    "paths": { 
+    "lib": ["dom", "dom.iterable", "esnext"],
+    "paths": {
       "@src/*": ["src/*"],
-      "@tests/*": ["src/__tests__/*"],
-      "@pages/*": ["src/pages/*"],
-      "@components": ["src/components/*"],
-      "@utils": ["src/utils/*"],
-      "@styles": ["src/styles/*"],
-      "@generated": ["src/generated/*"],
-      "@queries": ["src/queries/*"],
-      "@mutations": ["src/graphql/mutations/*"],
-      "@fragments": ["src/graphql/fragments/*"],
+      "@common/*": ["../abb-common/src/*"]
     },
     "allowJs": true,
     "skipLibCheck": true,
@@ -26,7 +14,6 @@
     "strictFunctionTypes": true,
     "noImplicitThis": true,
     "noUnusedLocals": true,
-    "noUnusedParameters": true,
     "noImplicitReturns": true,
     "noFallthroughCasesInSwitch": true,
     "forceConsistentCasingInFileNames": true,
@@ -37,7 +24,8 @@
     "resolveJsonModule": true,
     "isolatedModules": true,
     "jsx": "preserve",
-    "baseUrl": "./src"
+    "rootDir": "./src",
+    "baseUrl": "."
   },
   "include": [
     "next-env.d.ts",
